Allow non-equi joins if tables are alredy joined with an equi-join
We used to error out if the join clause includes filters like
t1.a < t2.a even if other filter like t1.key = t2.key exists.

Recently we lifted that restriction in subquery planning by
not lifting that restriction and focusing on equivalance classes
provided by postgres.

This checkin forwards previously erroring out real-time queries
due to join clauses to subquery planner and let it handle the
join even if the query does not have a subquery.

Conflicts:
	src/backend/distributed/planner/multi_logical_planner.c
	src/test/regress/expected/with_basics.out